Force the timeout value to integer after multiplying by 1000 on Windows. This lets us use floating point timeout values without error.
git-svn-id: svn+ssh://svn.dlogics.com/svn/dev/Packages/python_processes/trunk@22963 a347f3a1-5518-0410-bf23-827ba1738e83

+++ killableprocess.py
@@ -190,7 +190,7 @@ class Popen(subprocess.Popen):
 
         if mswindows:
             if timeout != -1:
-                timeout = timeout * 1000
+                timeout = int(timeout * 1000)
             rc = winprocess.WaitForSingleObject(self._handle, timeout)
             if rc == winprocess.WAIT_TIMEOUT:
                 self.kill(group)
